+++ assets/javascripts/kitten/components/search/search-input/index.js
@@ -29,4 +29,11 @@ export const SearchInput = styled.input.attrs({ type: 'search' })`
   ::placeholder {
     color: ${COLORS.line1};
   }
+
+  &::-webkit-search-decoration,
+  &::-webkit-search-cancel-button,
+  &::-webkit-search-results-button,
+  &::-webkit-search-results-decoration {
+    display: none;
+  }
 `