+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Added `Rocket` illustration.
+- Page: New `Thanks` page
 
 ## [1.7.0] - 2019-01-22